+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/ca/lsp/cobol/service/MyTextDocumentService.java
@@ -238,7 +238,7 @@ public class MyTextDocumentService implements TextDocumentService, EventObserver
     runAsync(
             () -> {
               AnalysisResult result =
-                  engine.analyze(uri, text, getCopybookProcessingMode(uri, DID_OPEN.name()));
+                  engine.analyze(uri, text, getCopybookProcessingMode(uri, DID_OPEN));
               ofNullable(docs.get(uri)).ifPresent(doc -> doc.setAnalysisResult(result));
               publishResult(uri, result);
             })
@@ -249,7 +249,7 @@ public class MyTextDocumentService implements TextDocumentService, EventObserver
     runAsync(
             () -> {
               AnalysisResult result =
-                  engine.analyze(uri, text, getCopybookProcessingMode(uri, DID_CHANGE.name()));
+                  engine.analyze(uri, text, getCopybookProcessingMode(uri, DID_CHANGE));
               registerDocument(uri, new MyDocumentModel(text, result));
               communications.publishDiagnostics(result.getDiagnostics());
             })