+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/usecases/TestCopybookNotProcessedInExtendedSource.java
@@ -15,7 +15,6 @@
 
 package com.ca.lsp.cobol.usecases;
 
-import com.ca.lsp.cobol.positive.CobolText;
 import com.ca.lsp.cobol.usecases.engine.UseCaseEngine;
 import org.junit.jupiter.api.Test;
 
@@ -24,7 +23,10 @@ import java.util.Map;
 
 import static com.ca.lsp.cobol.service.CopybookProcessingMode.DISABLED;
 
-/** */
+/**
+ * This test checks that the copybook not processed if this document is an extended document. Here:
+ * check the copybook not recognized and its content not resolved.
+ */
 class TestCopybookNotProcessedInExtendedSource {
   private static final String TEXT =
       "        IDENTIFICATION DIVISION.\r\n"